Deleted pyc file. Cleaning up of shared_functions.py, shared_variables.py, and acquire_and_preprocess_inputs.py for testing on VM.

+++ lib/utils/shared_functions.py
@@ -45,26 +45,31 @@ def subset_wbd_gpkg(args):
     
     # Parse geopackage path.
     wbd_gpkg = args[0]
+    wbd_gpkg_conus_output = args[1]
     
     print("Subsetting " + wbd_gpkg + "...")
     
     # Read geopackage into dataframe.
     wbd = gp.read_file(wbd_gpkg)
     gdf=gp.GeoDataFrame(wbd)
-    
-    # Delete OCONUS polygons from the watershed boundary dataset.
+        
     for index, row in gdf.iterrows():
         state = row["STATES"]
-        wbd_state_list = state.split(",")  # Some polygons have multiple states, separated by a comma.
-        keep_flag = False  # Default to Fault, i.e. to delete the polygon.
-        for wbd_state in wbd_state_list:
-            if wbd_state in CONUS_STATE_LIST:  # Check each polygon to make sure it's state abbrev name is allowed.
+        if state != None:  # Some polygons are empty in the STATES field.
+            keep_flag = False  # Default to Fault, i.e. to delete the polygon.
+            if state in CONUS_STATE_LIST:
                 keep_flag = True
-        if not keep_flag:
-            gdf.drop(index, inplace=True)  # Delete from dataframe.
-        
-        # Overwrite geopackage.
-        gdf.to_file(wbd_gpkg, driver='GPKG')
+            # Only split if multiple states present. More efficient this way.
+            elif len(state) > 2:
+                for wbd_state in state.split(","):  # Some polygons have multiple states, separated by a comma.
+                    if wbd_state in CONUS_STATE_LIST:  # Check each polygon to make sure it's state abbrev name is allowed.
+                        keep_flag = True
+                        break
+            if not keep_flag:
+                gdf.drop(index, inplace=True)  # Delete from dataframe.
+
+    # Overwrite geopackage.
+    gdf.to_file(wbd_gpkg_conus_output, driver='GPKG')
